requirements.txt
allure-pytest==2.5.4
allure-python-commons==2.5.4
atomicwrites==1.3.0
attrs==19.1.0
more-itertools==6.0.0
pluggy==0.9.0
py==1.8.0
pytest==4.3.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
pytest-xdist==1.26.1
requests==2.21.0
optimize-images==1.3
Pillow==5.4.1
